refactor(ack,mllp-ack)!: ack is the acknowledgment language; mllp-ack owns responses
@glion/ack trims to typed vocabulary both sides of a connection share:
Table 0008/0357/0516 codes as merged const/type names, guards, and the
AckException family carrying data (errorCode, severity, text, controlId)
— with @glion/ast as its only dependency. Everything else leaves:

- acknowledge() moves to @glion/mllp-ack, exported beside the middleware;
  NAKs carry MSH + MSA only unless the app supplies the new errSegment
  hook (the ERR layout is HL7v2-version dependent: ELD in ERR-1 before
  v2.5, ERR-3/ERR-4 after — a version-agnostic package cannot render it)
- toErrSegment() and includeErrSegment are removed
- ackExceptionFor folds into its one consumer: mllp-client maps MSA-1
  with a compiler-checked exhaustive switch
- AckException.raw/tree become text (MSA-3): full HL7v2 payloads on
  exceptions flow into logs and error trackers — a PHI hazard — and had
  one consumer, which only wanted MSA-3
- AckException.code narrows to AckNakCode; the CLI cast disappears
- uid() is gone (superseded by @glion/util-uid, already on main)
- tests mirror the source tree 1:1; README rewritten reader-first
