+++ packages/mllp-ack/src/index.ts
@@ -1,1 +1,3 @@
+export { acknowledge } from "./acknowledge";
+export type { AcknowledgeOptions, SendingInfo } from "./acknowledge";
 export { ackMiddleware } from "./ack";
